Set the row format in the default table options (see #1221)
Description
-----------

InnoDB large prefixes require the row format to be `DYNAMIC`. Older MySQL and MariaDB versions set the row format to `COMPACT` by default, which is why we dynamically change it in the `DcaSchemaProvider::appendToSchema()` method.

https://github.com/contao/contao/blob/7f092db0b74fe289da64c8153bb4de0ca87edaab/core-bundle/src/Doctrine/Schema/DcaSchemaProvider.php#L83-L86

The problem with this workaround is that it does not work for ORM entities, from which we now have four. This causes an error during DB migration:

`An exception occurred while executing 'CREATE TABLE tl_cron_job (id INT UNSIGNED AUTO_INCREMENT NOT NULL, name VARCHAR(255) NOT NULL, lastRun DATETIME NOT   
  NULL, INDEX name (name), PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB': SQLSTATE[HY000]: General error: 1709 Index column size too large. The maximum column size is 767 bytes. `

This PR removes the hacky workaround and adds a configuration check in the install tool instead:

<img width="793" alt="" src="https://user-images.githubusercontent.com/1192057/72716114-f7071080-3b71-11ea-97e4-cde3d0cc5c7d.png">

If the row format is properly configured in the default table options, it works with both ORM entities and DCA tables.

Commits
-------

10797b42 Set the row format in the default table options
a16ccf14 Unset the row_format parameter in the MyISAM configuration example