Switch to Symfony's version of the Path helper (see #3695)
Description
-----------

The `webmozart/path-util` package got abandoned but the `Path` helper class was added to the `Filesystem` component of Symfony 5.4. 

Besides the `getFilename()` method removed (which is basically `basename`) and the use of a variadic instead of an array parameter in `getLongestCommonBasePath()` everything stayed the same.

The CI jobs must be rerun once Symfony 5.4 is released.

Commits
-------

4918c5ac switch to symfony's version of the path util class
9f55a46b adjust Webmozart\PathUtil usage in more classes
04da5c4d Merge branch '4.x' into bugfix/pathutil
35624068 Fix two left-overs