Set DB server version in install tool (see #4158)
Description
-----------

To reduce the likelihood of issues with the Doctrine server detection mechanism and to improve performance we should IMO always set the MySQL server version if possible.

Not sure if this needs/should get a backport for Contao 4.9.

Commits
-------

d61e1de3 Set DB server version in install tool